feat: continue adding to Lua port
Yeah, this is a horrible mixed-bag commit. Just showing that we can
compile multiple translation units in a single invocation (may want to
refactor the Makefile to do that a little more traditionally, to be
honest; compile each source individually, then link... but then again,
this is simple and the project is tiny so speed is not a concern).